Adv_Autonomous_Agent/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
pandas
openpyxl
python-pptx
uvicorn
fastapi
pydantic
reportlab
requests
python-dotenv
python-docx
openai
